+++ main.go
@@ -85,6 +85,7 @@ func main() {
 	commands.Init_Install_Command(rootCmd, &app)
 	commands.Init_List_Command(rootCmd, &app)
 	commands.Init_Make_Command(rootCmd, &app)
+	commands.Init_Monitor_Command(rootCmd, &app)
 	commands.Init_New_Command(rootCmd, &app)
 	commands.Init_Open_Command(rootCmd, &app)
 	commands.Init_Pack_Command(rootCmd, &app)